refactor: Request Permission を修正
